Added AuthenticationErrors if the user/userrole is not able to perfom the action

+++ server/commands/UserCommand/DeleteUserCommand.js
@@ -1,4 +1,6 @@
 const MongoDBAdapter = require('../../databaseAdapters/mongoDB/MongoDBAdapter');
+const AuthenticationError = require('../../errors/AuthenticationError');
+const Userrole = require('../../dataModels/Userrole');
 const MainLogger = require('../../Logger.js').logger;
 
 const logger = MainLogger.child({ service: 'DeleteUserroleCommand' });
@@ -13,8 +15,16 @@ module.exports = {
   register(db) {
     database = db;
   },
-  async execute(username) {
+  async execute(usernameActionPerformedBy, username) {
     logger.info('Executing DeleteUserCommand with username: %s', username);
+
+    const dbUserActionPerfomedBy = await database.getUser(usernameActionPerformedBy);
+    const dbUserActionPerfomedTo = await database.getUser(username);
+
+    if (!dbUserActionPerfomedBy.userrole.rules.canEditUser.allowed
+      || !await Userrole.canUserroleEditUserrole(dbUserActionPerfomedBy.userrole.name, dbUserActionPerfomedTo.userrole.name, database)) {
+      throw new AuthenticationError(dbUserActionPerfomedBy.userrole.name, `Userrole ${dbUserActionPerfomedBy.userrole.name} can not delete user ${username}`);
+    }
 
     return database.deleteUser(username);
   },
